refactor: API 파일 통합
