--- components/Api/ApiErrorAlert.tsx
@@ -1,10 +1,0 @@
-import { Alert, AlertIcon } from '@chakra-ui/react';
-
-export default function ApiErrorAlert({ error }: any) {
-  return (
-    <Alert status="error">
-      <AlertIcon />
-      {error}
-    </Alert>
-  );
-}
